Moved Platform class to data package
