--- shared/src/commonMain/kotlin/com/vickikbt/kmptemplate/presentation/utils/Greeting.kt
@@ -1,9 +1,0 @@
-package com.vickikbt.kmptemplate.presentation.utils
-
-import com.vickikbt.kmptemplate.domain.utils.Platform
-
-class Greeting {
-    fun greet(): String {
-        return "Hello, ${Platform().platform}!"
-    }
-}
