--- shared/src/commonMain/kotlin/com/vickikbt/kmptemplate/presentation/viewmodels/MainViewModel.kt
@@ -1,25 +1,0 @@
-package com.vickikbt.kmptemplate.presentation.viewmodels
-
-import com.vickikbt.kmptemplate.presentation.utils.Greeting
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.launch
-import org.koin.core.component.KoinComponent
-
-class MainViewModel constructor(private val greetingRepository: Greeting) : KoinComponent {
-
-    // @NativeCoroutineScope
-    private val viewModelScope = CoroutineScope(Dispatchers.Default)
-
-    private val _greeting = MutableStateFlow<String?>(null)
-    val greeting get() = _greeting
-
-    init {
-        greetings()
-    }
-
-    private fun greetings() = viewModelScope.launch {
-        _greeting.value = greetingRepository.greet()
-    }
-}
