+++ shared/src/jvmMain/kotlin/com/vickikbt/kmptemplate/data/Platform.kt
@@ -1,4 +1,4 @@
-package com.vickikbt.kmptemplate.domain.utils
+package com.vickikbt.kmptemplate.data
 
 actual class Platform actual constructor() {
     actual val platform: String = "Desktop"
